+++ examples/cns_2d_shock_tube/clean.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+rm -f field0.* user.mod user.smod *.log neko