+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Develop
 
+- Add physical viscous and conductive flux support to the compressible solver.
 - Fixed stale facet-normals in fluid_pnpn pressure surface terms.
   This was only affecting ALE simulations containing rotations.
 - *BREAKING* Changed the size of mesh velocity lag arrays in ALE to 2.